refactor(models): all mongoose models slight refactor;
